set fp16 prec in ai visual generator

+++ ai_ref_kits/multimodal_ai_visual_generator/ci/test.py
@@ -24,7 +24,7 @@ from convert_and_optimize_text2image import convert_image_model
 MODEL_DIR = Path("models")
 LLM_MODEL_TYPE = "tiny-llama-1b-chat"
 IMAGE_MODEL_TYPE = "lcm"
-PRECISION = "int4"
+PRECISION = "fp16"
 LOG_FILE = Path("gradio_log.txt")
 
 # ----- Step 1: Export Models if Needed (will handle download internally) -----
